// Assets/templates/fuse-makeithome/src/app/modules/admin/apps/notes/notes.types.ts
export interface Task
{
    id?: string;
    content?: string;
    completed?: string | boolean;
}

export interface Label
{
    id?: string;
    title?: string;
}

export interface Note
{
    id?: string;
    title?: string;
    content?: string;
    tasks?: Task[];
    image?: string | null;
    labels?: Label[];
    archived?: boolean;
    createdAt?: string;
    updatedAt?: string | null;
}